ParamaShop/parama_shopping/backend/requirements.txt
alembic==1.18.1
asgiref==3.11.0
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
Django==6.0.1
Flask==3.1.2
Flask-Cors==4.0.2
Flask-JWT-Extended==4.7.1
Flask-Mail==0.9.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.10.1
python-dotenv==1.2.1
SQLAlchemy==2.0.46
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.5
